Add new unit tests JobParameterParserTest, ConfigurationTest, SettingsTest.
The primary motivation of this commit is to thoroughly cover the behavior of the JobParameterParser, so it can be refactored without breaking existing functionality. Other tests are less developed, but give a starting point for more thorough testing.

Test classes were decided based on unit test coverage of existing code.